+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [26.1.2] 2020-10-05
+### Added
+- Support for using a FHIR Questionnaire as a form definition in addFormToPage.
+
 ## [26.1.1] 2020-09-30
 ### Fixed
 - Fixed the extensions URI for answer scores